Delete slack.sls
This Slack installer is not optimized for deployments, since it assumes that the installation is done by a windows local or domain user. 
In most cases, the salt-minion service runs as a system or special user, the state always installs the application at the user specific profile.
In the case of the system user on "C:\Windows\SysWOW64\config\systemprofile\AppData\Local\slack".