tortoise-hg: official download site and current version (5.7.1)
The mentioned BitBucket page is long gone. According to previous
comment, it looks as it had a technical difficulty (passing AWS S3
token, even for public download), which the current official download
page hasn't, so we don't see any reason to require pre-downloading the
MSI to the Salt fileserver for recent versions.

The proposed versions were very outdated, as Mercurial 3.6.2 was
released on 2015-11-24.