tortoise-hg: disabling install of versions 3.x.y
These versions have been unsupported by upstream for more than
four years. At this point, it is reasonable to consider them
a security liability by default.

The installer is not available anymore, but some
users may still have in their cache directory, since that was
the expected way of doing.